+++ lib/pymedphys/_experimental/autosegmentation/pipeline.py
@@ -19,7 +19,7 @@ import pathlib
 
 from pymedphys._imports import numpy as np
 from pymedphys._imports import pydicom
-from pymedphys._imports import tensorflow as tf
+from pymedphys._imports.slow import tensorflow as tf
 
 from pymedphys._data import download
