+++ lib/pymedphys/_imports/slow/__init__.py
@@ -7,7 +7,7 @@ HERE = pathlib.Path(__file__).parent
 IMPORTABLES = _magic.import_magic(HERE)
 
 # This will never actually run, but it helps pylint know what's going on
-if "numpy" not in IMPORTABLES:
+if "tensorflow" not in IMPORTABLES:
     from .imports import *
 
     raise ValueError("This section of code should never run")
