+++ lib/pymedphys/_imports/slow/imports.py
@@ -1,0 +1,3 @@
+# pylint: disable = unused-import, reimported, import-error
+
+import tensorflow
